Staking

+++ models/node.go
@@ -28,7 +28,7 @@ type Node struct {
 	Status                  NodeStatus     `json:"status" gorm:"index"`
 	GPUName                 string         `json:"gpu_name" gorm:"index"`
 	GPUVram                 uint64         `json:"gpu_vram" gorm:"index"`
-	QOSScore                uint64         `json:"qos_score"`
+	QOSScore                float64        `json:"qos_score"`
 	MajorVersion            uint64         `json:"major_version"`
 	MinorVersion            uint64         `json:"minor_version"`
 	PatchVersion            uint64         `json:"patch_version"`
